Experiment: else ifs in serializer code
Use scalameta to transform lists of mutually exclusive ifs into else ifs, to potentially save a miniscule amount of performance.
